[Decomp][ObjectMgr] Cohesion split: extract creature loaders to ObjectMgrCreatures.cpp
Moves 24 creature-related methods (~1,250 lines, contiguous range
lines 572..1811 in the previous file) into a new
src/game/Object/ObjectMgrCreatures.cpp:

  - LoadCreatureTemplates, LoadCreatureLocales-already-stays
  - ConvertCreatureAddonAuras, LoadCreatureAddons (x2)
  - LoadCreatureClassLvlStats, GetCreatureClassLvlStats
  - LoadEquipmentTemplates
  - GetCreatureModelAlternativeModel, GetCreatureModelRandomGender,
    GetModelForRace, LoadCreatureModelInfo, LoadCreatureModelRace
  - LoadCreatures, AddCreatureToGrid, RemoveCreatureFromGrid

Same `ObjectMgr` class -- this is a cohesion split, not an
extraction. No header change, no caller change, no behaviour
change. ObjectMgr.h is untouched; ObjectMgr.cpp drops from
12,200 -> 10,957 lines.

CMakeLists.txt is unchanged: `file(GLOB Object/*.cpp)` picks up
the new file automatically.

First in a planned series of cohesion splits for ObjectMgr per
the MANGOS_SYSTEMS_AUDIT.md decomposition roadmap. Subsequent
PRs will split out items, players, quests, gameobjects, world
data, etc. Each PR keeps the same single-class single-header
layout and only moves method bodies to a new .cpp.

Antz pre-approved the broader decomposition arc per the
"god-object decomposition" sync; mangostwo not synced because
this is purely a mangosthree source-tree organisation change
(the file name and method bodies are unchanged).
